+++ Benchmarks.cs
@@ -1100,12 +1100,12 @@ public class Benchmarks : JobComponentSystem {
 			return value;
 		}
 
-		private ulong Compute(byte* buffer, int length, ulong a, ulong b, ulong c, ulong d) {
-			const int blockSize = 32;
+		private ulong Compute(byte* buffer, ulong length, ulong a, ulong b, ulong c, ulong d) {
+			const uint blockSize = 32;
 
-			int end = length & ~(blockSize - 1);
+			ulong end = length & ~(blockSize - 1);
 
-			for (int i = 0; i < end; i += blockSize) {
+			for (uint i = 0; i < end; i += blockSize) {
 				a ^= Read(buffer + i);
 				b ^= Read(buffer + i + 8);
 				c ^= Read(buffer + i + 16);
@@ -1117,7 +1117,7 @@ public class Benchmarks : JobComponentSystem {
 				d = Diffuse(d);
 			}
 
-			int excessive = length - end;
+			ulong excessive = length - end;
 			byte* bufferEnd = buffer + end;
 
 			if (excessive > 0) {
@@ -1146,7 +1146,7 @@ public class Benchmarks : JobComponentSystem {
 			a ^= b;
 			c ^= d;
 			a ^= c;
-			a ^= (ulong)length;
+			a ^= length;
 
 			return Diffuse(a);
 		}
